# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Traffic Digital Twin API",
    "version": "1.0.0",
    "description": "Local functional engineering prototype: authenticated Go control and durable evidence, recorded aggregate observations, causal forecasts and virtual signal comparisons. Acceptance remains open; physical signal control and field readiness are disabled/deferred."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8081"
    }
  ],
  "paths": {
    "/api/v1/network": {
      "get": {
        "summary": "Validated graph and signal configuration",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Network"
                }
              }
            }
          },
          "503": {
            "description": "Component unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/state": {
      "get": {
        "summary": "Latest validated traffic state; unavailable until simulation is connected",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrafficState"
                }
              }
            }
          },
          "503": {
            "description": "Component unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/api/v1/junctions/{id}": {
      "get": {
        "summary": "Configured node and phases; no current measurements",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "503": {
            "description": "Component unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": []
      }
    },
    "/api/v1/runs": {
      "get": {
        "summary": "Latest 50 persisted runs",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Run"
                  }
                }
              }
            }
          },
          "503": {
            "description": "Component unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Verified unexpired server session required"
          }
        }
      },
      "post": {
        "summary": "Prepare a run and atomic audit record; does not start simulation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCommand"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Run persisted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "400": {
            "description": "Invalid command",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Cross-origin command denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Persistence unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Verified unexpired server session required"
          }
        }
      }
    },
    "/api/v1/audit": {
      "get": {
        "summary": "Append-only audit history in ascending sequence order",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "503": {
            "description": "Component unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Verified unexpired server session required"
          }
        },
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ]
      }
    },
    "/api/v1/recommendations/active": {
      "get": {
        "summary": "Active recommendation; unavailable in manual mode or on service failure",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recommendation"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "409": {
            "description": "Stale, locked, unsafe or already decided"
          },
          "503": {
            "description": "Dependency unavailable; inspect audit if application was not confirmed"
          },
          "401": {
            "description": "Verified unexpired server session required"
          }
        }
      }
    },
...
    "/api/v1/scenarios/{type}/start": {
      "post": {
        "summary": "Start a seeded aggregate-flow scenario and persist activation/audit",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid request"
          },
          "409": {
            "description": "Stale, locked, unsafe or already decided"
          },
          "503": {
            "description": "Dependency unavailable; inspect audit if application was not confirmed"
          },
          "401": {
            "description": "Verified unexpired server session required"
          },
          "403": {
            "description": "Matching Origin and authenticated role required; viewers cannot mutate"
          }
        },
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "peak_surge",
                "incident_c3",
                "ambulance_corridor"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "schema_version",
                  "seed",
                  "mode"
                ],
                "properties": {
                  "schema_version": {
                    "const": "1.0"
                  },
                  "seed": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 4294967295
                  },
                  "mode": {
                    "enum": [
                      "observe",
                      "recommend",
                      "manual"
                    ]
                  },
                  "incident": {
                    "type": "object",
                    "description": "Optional virtual C3 incident override; valid only when type is incident_c3.",
                    "additionalProperties": false,
                    "required": [
                      "kind",
                      "capacity_ratio"
                    ],
                    "properties": {
                      "kind": {
                        "const": "capacity_reduction"
                      },
                      "capacity_ratio": {
                        "type": "number",
                        "minimum": 0.1,
                        "maximum": 0.9
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scenarios/reset": {
      "post": {
        "summary": "Repeat the current scenario and seed with a new audited run ID",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid request"
          },
          "409": {
            "description": "Stale, locked, unsafe or already decided"
          },
          "503": {
            "description": "Dependency unavailable; inspect audit if application was not confirmed"
          },
          "401": {
            "description": "Verified unexpired server session required"
          },
          "403": {
            "description": "Matching Origin and authenticated role required; viewers cannot mutate"
          }
        }
      }
    },
    "/api/v1/analysis": {
      "get": {
        "summary": "Fresh forecasts, recommendation, alternatives and equal-state comparison",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Analysis"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "409": {
            "description": "Stale, locked, unsafe or already decided"
          },
          "503": {
            "description": "Dependency unavailable; inspect audit if application was not confirmed"
          },
          "401": {
            "description": "Verified unexpired server session required"
          }
        }
      }
    },
    "/api/v1/replay/{scenario}": {
      "post": {
        "summary": "Start a checksummed eight-minute recording through the same state/WebSocket pipeline; actuation disabled",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid request"
          },
          "409": {
            "description": "Stale, locked, unsafe or already decided"
          },
          "503": {
            "description": "Dependency unavailable; inspect audit if application was not confirmed"
          },
          "401": {
            "description": "Verified unexpired server session required"
          },
          "403": {
            "description": "Matching Origin and authenticated role required; viewers cannot mutate"
          }
        },
        "parameters": [
          {
            "name": "scenario",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "peak_surge",
                "incident_c3",
                "ambulance_corridor"
              ]
            }
          }
        ]
      }
    },
    "/api/v1/vision/{id}": {
      "get": {
        "summary": "Get sample video traffic aggregates, tracking states and upstream impacts for a camera-monitored intersection (C3)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Intersection identifier (e.g. C3)"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Force simulated offline state (e.g. status=offline)"
          }
        ],
        "responses": {
          "200": {
            "description": "Traffic aggregates from computer vision processing or explicit unavailable status"
          },
          "404": {
            "description": "No vision camera configured for the requested junction"
          },
          "503": {
            "description": "Vision service offline or dependency unavailable"
          },
          "401": {
            "description": "Verified unexpired server session required"
          }
        }
      }
    },
    "/api/v1/commands/{id}": {
      "get": {
        "summary": "Retrieve recorded outcome of an uncertain or completed idempotent command",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Idempotency command identifier (8-128 chars)"
          }
        ],
        "responses": {
          "200": {
            "description": "Recorded command outcome"
          },
          "400": {
            "description": "Invalid command ID format"
          },
          "404": {
            "description": "Command not found"
          },
          "409": {
            "description": "Command actor or payload conflict"
          },
          "401": {
            "description": "Verified unexpired server session required"
          }
        }
      }
    },
    "/api/v1/session": {
      "get": {
        "summary": "Get current authenticated session actor and role",
        "responses": {
          "200": {
            "description": "Current session information"
          },
          "401": {
            "description": "Verified unexpired server session required"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "RATE_LIMITED",
              "INTELLIGENCE_TIMEOUT",
              "SIMULATION_UNAVAILABLE",
              "DATABASE_UNAVAILABLE",
              "INTERNAL_ERROR"
            ]
          },
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "twin_session",
        "description": "Expiring and revocable server session; actor and role derived server-side. Mutation requires matching Origin."
      }
    }
  },
  "security": [
    {
      "SessionCookie": []
    }
  ]
}